Tests for quote qty

+++ src/ExchangeBalanceComponent.cairo
@@ -165,7 +165,7 @@ mod exchange_balance_logic_component {
             self.emit(Transfer{from_:from, to, token, amount});
         }
 
-        fn validate_and_apply_gas_fee_internal(ref self: ComponentState<TContractState>, user: ContractAddress, gas_fee: super::GasFee, gas_price: u256, times:u8, cur_gas_per_action:u32) -> u256 {
+        fn validate_and_apply_gas_fee_internal(ref self: ComponentState<TContractState>, user: ContractAddress, gas_fee: super::GasFee, gas_price: u256, times:u16, cur_gas_per_action:u32) -> u256 {
             // Validates that gas_fee correctly specified, calculate how many coins and in what token we tfer from user to exchange  
             if gas_price == 0 || gas_fee.gas_per_action == 0 { return 0;}
             let (spent, coin) = super::get_gas_fee_and_coin(gas_fee, gas_price, self.wrapped_native_token.read(), cur_gas_per_action);
